# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Hybrid Search API
  description: |
    Conversational Hybrid Search & RAG System for Catering Menus

    This API provides a multi-tenant platform for restaurant menu search, ordering, and payment processing.
    It combines BM25 lexical search, vector similarity search, and optional graph-based search to deliver
    intelligent conversational search results.

    ## Key Features
    - **Hybrid Search**: Combines BM25, vector (pgvector), and graph (Neo4j) search with RRF fusion
    - **Conversational AI**: Session-based conversational search with intent detection
    - **Multi-tenancy**: Full tenant isolation with Row-Level Security (RLS)
    - **Authentication**: JWT-based auth with role-based access control (RBAC)
    - **Payments**: Stripe Connect integration for restaurant payments
    - **POS Integration**: Square, Toast, and other POS system integrations
    - **Order Management**: Complete order lifecycle from cart to fulfillment

    ## Feature Flags
    - `ENABLE_GRAPH_SEARCH`: Enable Neo4j graph search (default: false)
    - `ENABLE_3WAY_RRF`: Enable 3-way RRF fusion (BM25 + vector + graph) (default: false)
    - `ENABLE_PAYMENTS`: Enable Stripe payment processing (default: false)
    - `ENABLE_POS_INTEGRATION`: Enable POS system integrations (default: false)

    ## Authentication
    Most endpoints require JWT Bearer authentication. Obtain tokens via `/auth/register` or `/auth/login`.
...

  version: 0.1.0
  contact:
...
  license:
...
servers:
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login or /auth/register
  schemas:
...
    SearchRequest:
      type: object
      required:
        - session_id
        - user_input
      properties:
        session_id:
          type: string
          minLength: 8
          maxLength: 64
          description: Session ID for conversational context
          example: "session_abc123xyz789"
        user_input:
          type: string
          maxLength: 500
          description: User's search query or conversational input
          example: "Find Italian catering in Boston for 20 people"
        max_results:
          type: integer
          minimum: 1
          maximum: 50
          default: 10
          description: Maximum number of results to return
    MenuItemResult:
      type: object
      required:
        - doc_id
        - restaurant_name
        - city
        - state
        - item_name
      properties:
        doc_id:
          type: string
          description: Unique document ID
          example: "item_123"
        restaurant_name:
          type: string
          description: Restaurant name
          example: "Tony's Catering"
        city:
          type: string
          example: "Boston"
        state:
          type: string
          example: "MA"
        item_name:
          type: string
          description: Menu item name
          example: "Italian Wedding Buffet"
        item_description:
          type: string
          nullable: true
          description: Detailed item description
          example: "Classic Italian buffet with pasta, salad, and bread"
        display_price:
          type: number
          format: float
          nullable: true
          description: Display price (may be per person or total)
          example: 15.99
        price_per_person:
          type: number
          format: float
          nullable: true
          description: Price per person if applicable
          example: 15.99
        serves_min:
          type: integer
          nullable: true
          description: Minimum number of people served
          example: 10
        serves_max:
          type: integer
          nullable: true
          description: Maximum number of people served
          example: 50
        dietary_labels:
          type: array
          items:
            type: string
          description: Dietary labels (vegetarian, vegan, gluten-free, etc.)
          example: ["vegetarian", "gluten-free"]
        tags:
          type: array
          items:
            type: string
          description: Additional tags for categorization
          example: ["italian", "buffet", "catering"]
        rrf_score:
          type: number
          format: float
          description: Reciprocal Rank Fusion score (0-1, higher is better)
          example: 0.856
    SearchResponse:
      type: object
      required:
        - session_id
        - resolved_query
        - intent
        - is_follow_up
        - filters
        - results
        - answer
        - confidence
        - processing_time_ms
      properties:
        session_id:
          type: string
          description: Session ID
          example: "session_abc123xyz789"
        resolved_query:
          type: string
          description: Query after context resolution and rewriting
          example: "Italian catering restaurants in Boston serving 20 people"
        intent:
          type: string
          enum: [search, filter, clarify, compare, graph]
          description: Detected user intent
          example: "search"
        is_follow_up:
          type: boolean
          description: Whether this is a follow-up query
          example: false
        filters:
          type: object
          additionalProperties: true
          description: Extracted filters (city, cuisine, price, etc.)
          example:
            city: "Boston"
            cuisine: "Italian"
            min_people: 20
        results:
          type: array
          items:
            $ref: '#/components/schemas/MenuItemResult'
          description: Search results
        answer:
          type: string
          description: Natural language answer generated by RAG
          example: "I found 8 Italian catering options in Boston that can serve 20 people..."
        confidence:
          type: number
          format: float
          description: Confidence score (0-1)
          example: 0.92
        processing_time_ms:
          type: number
          format: float
          description: Processing time in milliseconds
          example: 234.56
    SessionResponse:
      type: object
      required:
        - session_id
        - created_at
        - last_activity
        - entities
        - conversation_length
        - previous_results_count
      properties:
        session_id:
          type: string
          example: "session_abc123xyz789"
        created_at:
          type: string
          format: date-time
          description: Session creation timestamp
        last_activity:
          type: string
          format: date-time
          description: Last activity timestamp
        entities:
          type: object
          additionalProperties: true
          description: Extracted entities (locations, cuisines, etc.)
          example:
            city: "Boston"
            cuisine: "Italian"
        conversation_length:
          type: integer
          description: Number of conversation turns
          example: 3
        previous_results_count:
          type: integer
          description: Number of results from previous search
          example: 8
    FeedbackRequest:
      type: object
      required:
        - doc_id
        - rating
      properties:
        doc_id:
          type: string
          description: Document ID being rated
          example: "item_123"
        rating:
          type: integer
          minimum: 1
          maximum: 5
          description: User rating (1-5 stars)
          example: 5
        comment:
          type: string
          nullable: true
          description: Optional feedback comment
          example: "Exactly what I was looking for!"

    # Authentication Models
...
    CreateTenantRequest:
      type: object
      required:
        - name
        - slug
        - contact_email
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Tenant name
          example: "Acme Catering Co"
        slug:
          type: string
          minLength: 1
          maxLength: 255
          pattern: '^[a-z0-9-]+$'
          description: URL-friendly slug (lowercase, alphanumeric, hyphens)
          example: "acme-catering"
        contact_email:
          type: string
          format: email
          description: Primary contact email
          example: "contact@acme-catering.com"
        contact_phone:
          type: string
          nullable: true
          example: "+1-555-123-4567"
        billing_email:
          type: string
          format: email
          nullable: true
          example: "billing@acme-catering.com"
        application_fee_percent:
          type: number
          format: float
          minimum: 0
          maximum: 30
          default: 2.5
          description: Platform fee percentage (0-30%)
          example: 2.5
...
    TenantResponse:
      type: object
      required:
        - id
        - name
        - slug
        - is_active
        - application_fee_percent
        - restaurant_count
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        is_active:
          type: boolean
        contact_email:
          type: string
          format: email
          nullable: true
        contact_phone:
          type: string
          nullable: true
        billing_email:
          type: string
          format: email
          nullable: true
        application_fee_percent:
          type: number
          format: float
          description: Platform fee percentage (e.g., 2.5 for 2.5%)
        restaurant_count:
          type: integer
          description: Number of restaurants under this tenant
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TenantListResponse:
      type: object
      required:
        - items
        - total
        - limit
        - offset
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TenantResponse'
        total:
          type: integer
          description: Total number of tenants matching filters
        limit:
          type: integer
        offset:
          type: integer

    # Restaurant Models
    CreateRestaurantRequest:
      type: object
      required:
        - name
        - slug
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          example: "Tony's Catering"
        slug:
          type: string
          minLength: 1
          maxLength: 255
          pattern: '^[a-z0-9-]+$'
          example: "tonys-catering"
        description:
          type: string
          nullable: true
          example: "Family-owned Italian catering since 1985"
        address_line1:
          type: string
          nullable: true
          example: "123 Main St"
        address_line2:
          type: string
          nullable: true
          example: "Suite 200"
        city:
          type: string
          nullable: true
          example: "Boston"
        state:
          type: string
          nullable: true
          example: "MA"
        postal_code:
          type: string
          nullable: true
          example: "02101"
        country:
          type: string
          maxLength: 2
          default: US
          example: "US"
        phone:
          type: string
          nullable: true
          example: "+1-555-123-4567"
        email:
          type: string
          format: email
          nullable: true
          example: "orders@tonyscatering.com"
        timezone:
          type: string
          default: America/New_York
          example: "America/New_York"
...
    RestaurantResponse:
      type: object
      required:
        - id
        - tenant_id
        - name
        - slug
        - is_active
        - country
        - timezone
        - is_accepting_orders
        - stripe_connected
        - pos_connected
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        is_active:
          type: boolean
        address_line1:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        postal_code:
          type: string
          nullable: true
        country:
          type: string
        phone:
          type: string
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        timezone:
          type: string
        is_accepting_orders:
          type: boolean
        stripe_connected:
          type: boolean
          description: Whether Stripe Connect is set up and enabled
        pos_connected:
          type: boolean
          description: Whether a POS system is connected
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message
          example: "Authentication required"
...
paths:
...
  /chat/search:
    post:
      tags:
        - Search
      summary: Execute conversational search
      description: |
        Execute a conversational hybrid search query. This endpoint:
        1. Loads or creates a session for conversational context
        2. Runs the LangGraph pipeline (intent detection, query rewriting, hybrid search, RAG)
        3. Updates session with results
        4. Returns search results with natural language answer

        Supports follow-up queries that reference previous context.

        Examples:
        - "Find Italian catering in Boston for 20 people"
        - "Show me vegetarian options" (follow-up)
        - "Cheaper ones under $15 per person" (follow-up)
      operationId: chatSearch
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
            examples:
              initial_query:
                summary: Initial search query
                value:
                  session_id: "session_abc123"
                  user_input: "Find Italian catering in Boston for 20 people"
                  max_results: 10
              follow_up:
                summary: Follow-up query
                value:
                  session_id: "session_abc123"
                  user_input: "Show me vegetarian options"
                  max_results: 10
      responses:
        '200':
          description: Search results with conversational answer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Session Endpoints
  /session/{session_id}:
    get:
      tags:
        - Sessions
      summary: Get session state
      description: Retrieve session state including conversation history and entities
      operationId: getSession
      security: []
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
          description: Session ID
      responses:
        '200':
          description: Session state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Sessions
      summary: Delete session
      description: Clear session and start fresh
      operationId: deleteSession
      security: []
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: deleted
                  session_id:
                    type: string
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{session_id}/feedback:
    post:
      tags:
        - Sessions
      summary: Submit relevance feedback
      description: Submit user feedback on search result quality (for model improvement)
      operationId: submitFeedback
      security: []
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackRequest'
      responses:
        '200':
          description: Feedback received
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: received
                  doc_id:
                    type: string
                  rating:
                    type: integer

  # Authentication Endpoints
...
  /tenants:
    post:
      tags:
        - Tenants
      summary: Create tenant
      description: Create a new tenant (platform admin only)
      operationId: createTenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantRequest'
      responses:
        '201':
          description: Tenant created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          description: Tenant slug already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Tenants
      summary: List tenants
      description: List all tenants (platform admin only)
      operationId: listTenants
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: is_active
          in: query
          schema:
            type: boolean
          description: Filter by active status
      responses:
        '200':
          description: List of tenants
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
  /restaurants:
    post:
      tags:
        - Restaurants
      summary: Create restaurant
      description: Create a new restaurant for the tenant (restaurant_admin or platform_admin)
      operationId: createRestaurant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRestaurantRequest'
      responses:
        '201':
          description: Restaurant created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestaurantResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          description: Restaurant slug already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Restaurants
      summary: List restaurants
      description: List restaurants for the tenant
      operationId: listRestaurants
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: List of restaurants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RestaurantResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...